Platform/Injectors: Fix logging bugs
Avoid buffer overflows in LogInjectionStatus and fix several Intel HD
Graphics injection bugs for older IGPUs.

+++ rEFIt_UEFI/Platform/Injectors/DeviceProperty.h
@@ -23,7 +23,7 @@ enum {
 
 typedef struct {
   CHAR16 *Name;
-  UINT32 Type;
+  UINT8  Type;
   VOID   *Value;
   UINT32 Size;
   UINT8  Flags;
